[PIPE-182] Moving spark conftest stuff to a separate module and adding remainder of tests for table creation
